Fix some rubocop failures

+++ app/services/plugins/entry/on_instance_destuction.rb
@@ -8,9 +8,7 @@ module Plugins
       end
 
       def call
-        if @info.addon_names.find { |addon_name| addon_name == "AWS S3" }
-          AwsIntegration::S3Accessor.new.delete_bucket(@info.application_name)
-        end
+        AwsIntegration::S3Accessor.new.delete_bucket(@info.application_name) if @info.addon_names.find { |addon_name| addon_name == "AWS S3" }
 
         ReturnValue.ok
       end
